OK-496: Siirretty Next.js-spesifit teema-koodit omaan tiedostoonsa
