+++ src/colors.tsx
@@ -1,0 +1,50 @@
+export const colors = {
+  grey900: '#1D1D1D',
+  grey800: '#454545',
+  grey700: '#4C4C4C',
+  grey600: '#5D5D5D',
+  grey500: '#6D6D6D',
+  grey400: '#B2B2B2',
+  grey300: '#C8C8C8',
+  grey200: '#DFDFDF',
+  grey100: '#EDEDED',
+  grey50: '#F6F6F6',
+
+  white: '#FFFFFF',
+  black: '#000000',
+
+  blue1: '#000066',
+  blue2: '#0033CC',
+  blue3: '#0041DC',
+  cyan1: '#006699',
+  cyan2: '#66CCCC',
+  cyan3: '#99FFFF',
+  lightBlue1: '#82D4FF',
+  lightBlue2: '#C1EAFF',
+
+  green1: '#254905',
+  green2: '#378703', // Success
+  green3: '#5BCA13',
+  green4: '#9CFF5A',
+  green5: '#CCFFCC',
+
+  red1: '#990066',
+  red2: '#E60895',
+
+  orange1: '#663300',
+  orange2: '#993300',
+  orange3: '#CC3300', // Error
+  orange4: '#FF5500',
+
+  yellow1: '#FFCC33',
+  yellow2: '#FFD900',
+  yellow3: '#FFFF33',
+
+  pink1: '#FF66CC',
+  pink2: '#FFCCFF',
+
+  purple1: '#660066',
+  purple2: '#660099',
+  purple3: '#C227B9',
+  purple4: '#CC99FF',
+};
