+++ src/theme.tsx
@@ -1,13 +1,14 @@
 'use client';
 
+import { colors } from './colors';
+import type { DeepPartial } from './util';
 import type { ButtonOwnProps } from '@mui/material';
 import {
   type Theme,
   type ThemeOptions,
   createTheme,
 } from '@mui/material/styles';
-import NextLink, { type LinkProps } from 'next/link';
-import React from 'react';
+import { deepmerge } from '@mui/utils';
 
 declare module '@mui/material/styles' {
   interface CustomTypographyVariants {
@@ -43,63 +44,6 @@ declare module '@mui/material/Typography' {
   }
 }
 
-const LinkBehaviour = React.forwardRef<HTMLAnchorElement, LinkProps>(
-  function LinkBehaviour(props, ref) {
-    return <NextLink ref={ref} {...props} />;
-  },
-);
-
-export const colors = {
-  grey900: '#1D1D1D',
-  grey800: '#454545',
-  grey700: '#4C4C4C',
-  grey600: '#5D5D5D',
-  grey500: '#6D6D6D',
-  grey400: '#B2B2B2',
-  grey300: '#C8C8C8',
-  grey200: '#DFDFDF',
-  grey100: '#EDEDED',
-  grey50: '#F6F6F6',
-
-  white: '#FFFFFF',
-  black: '#000000',
-
-  blue1: '#000066',
-  blue2: '#0033CC',
-  blue3: '#0041DC',
-  cyan1: '#006699',
-  cyan2: '#66CCCC',
-  cyan3: '#99FFFF',
-  lightBlue1: '#82D4FF',
-  lightBlue2: '#C1EAFF',
-
-  green1: '#254905',
-  green2: '#378703', // Success
-  green3: '#5BCA13',
-  green4: '#9CFF5A',
-  green5: '#CCFFCC',
-
-  red1: '#990066',
-  red2: '#E60895',
-  
-  orange1: '#663300',
-  orange2: '#993300',
-  orange3: '#CC3300', // Error
-  orange4: '#FF5500',
-
-  yellow1: '#FFCC33',
-  yellow2: '#FFD900',
-  yellow3: '#FFFF33',
-
-  pink1: '#FF66CC',
-  pink2: '#FFCCFF',
-
-  purple1: '#660066',
-  purple2: '#660099',
-  purple3: '#C227B9',
-  purple4: '#CC99FF',
-};
-
 const theme = createTheme({
   breakpoints: {
     values: {
@@ -113,9 +57,21 @@ const theme = createTheme({
   },
 });
 
+function getColorByName(
+  colorName: ButtonOwnProps['color'],
+  customTheme: Theme,
+  mode: 'main' | 'light' | 'dark',
+) {
+  return colorName === 'inherit'
+    ? 'inherit'
+    : customTheme.palette[colorName ?? 'primary'][mode];
+}
+
 const COMMON_THEME_OPTIONS: ThemeOptions = {
   typography: {
-    fontFamily: 'Open Sans',
+    allVariants: {
+      fontFamily: 'Open Sans, sans-serif',
+    },
     h1: {
       fontSize: '34px',
       fontWeight: 700,
@@ -184,11 +140,6 @@ const COMMON_THEME_OPTIONS: ThemeOptions = {
           subtitle1: 'body1',
           subtitle2: 'body2',
         },
-      },
-    },
-    MuiLink: {
-      defaultProps: {
-        component: LinkBehaviour,
       },
     },
     MuiButton: {
@@ -341,7 +292,7 @@ const COMMON_THEME_OPTIONS: ThemeOptions = {
   },
 };
 
-export const virkailijaTheme = createTheme({
+const VIRKAILIJA_THEME_OPTIONS = {
   ...COMMON_THEME_OPTIONS,
   palette: {
     primary: {
@@ -351,9 +302,9 @@ export const virkailijaTheme = createTheme({
       contrastText: colors.white,
     },
   },
-});
+};
 
-export const oppijaTheme = createTheme({
+const OPPIJA_THEME_OPTIONS = {
   ...COMMON_THEME_OPTIONS,
   palette: {
     primary: {
@@ -363,14 +314,26 @@ export const oppijaTheme = createTheme({
       contrastText: colors.white,
     },
   },
-});
+};
 
-function getColorByName(
-  colorName: ButtonOwnProps['color'],
-  customTheme: Theme,
-  mode: 'main' | 'light' | 'dark',
-) {
-  return colorName === 'inherit'
-    ? 'inherit'
-    : customTheme.palette[colorName ?? 'primary'][mode];
+type CreateOPHThemeParams = {
+  variant: 'oph' | 'opintopolku';
+  overrides?: DeepPartial<ThemeOptions>;
+};
+
+export function createODSTheme({
+  variant,
+  overrides = {},
+}: CreateOPHThemeParams) {
+  switch (variant) {
+    case 'oph':
+      return createTheme(deepmerge(VIRKAILIJA_THEME_OPTIONS, overrides));
+    case 'opintopolku':
+      return createTheme(deepmerge(OPPIJA_THEME_OPTIONS, overrides));
+    default:
+      throw Error('Theme variant must be "oph" or "opintopolku"!');
+  }
 }
+
+export const virkailijaTheme = createODSTheme({ variant: 'oph' });
+export const oppijaTheme = createODSTheme({ variant: 'opintopolku' });
